fix: update analytics dependencies to resolve vulnerabilities (#3976)
