+++ analytics/requirements.txt
@@ -1,165 +1,124 @@
-alabaster==0.7.13
--e "git+https://github.com/DataBiosphere/data-browser.git@60dc83eb36468c1f8091136f8e0534b3cba81b44#egg=analytics&subdirectory=analytics/analytics_package"
-anyio==3.6.2
+accessible-pygments==0.0.4
+alabaster==0.7.16
+-e git+https://github.com/DataBiosphere/data-browser.git@60dc83eb36468c1f8091136f8e0534b3cba81b44#egg=analytics&subdirectory=analytics/analytics_package
 appdirs==1.4.4
-appnope==0.1.3
-argon2-cffi==21.3.0
-argon2-cffi-bindings==21.2.0
-asttokens==2.2.1
-attrs==21.4.0
-Babel==2.11.0
-backcall==0.2.0
-beautifulsoup4==4.11.2
-bleach==6.0.0
-cachetools==5.3.0
-certifi==2022.12.7
-cffi==1.15.1
-charset-normalizer==3.0.1
-click==8.1.3
-colorama==0.4.6
-comm==0.1.2
-contourpy==1.0.7
-cycler==0.11.0
-db-dtypes==1.0.5
-debugpy==1.6.6
+appnope==0.1.4
+asttokens==2.4.1
+attrs==23.2.0
+Babel==2.14.0
+beautifulsoup4==4.12.3
+cachetools==5.3.3
+certifi==2024.2.2
+charset-normalizer==3.3.2
+click==8.1.7
+comm==0.2.2
+contourpy==1.2.1
+cycler==0.12.1
+debugpy==1.8.1
 decorator==5.1.1
-defusedxml==0.7.1
-docutils==0.17.1
-entrypoints==0.4
-executing==1.2.0
-fastjsonschema==2.16.2
-fonttools==4.38.0
-gitdb==4.0.10
-GitPython==3.1.30
-google-api-core==2.11.0
-google-api-python-client==2.77.0
-google-auth==2.16.0
-google-auth-httplib2==0.1.0
-google-auth-oauthlib==1.0.0
-google-cloud-bigquery==3.5.0
-google-cloud-core==2.3.2
-google-crc32c==1.5.0
-google-resumable-media==2.4.1
-googleapis-common-protos==1.58.0
-grpcio==1.51.1
-grpcio-status==1.51.1
-httplib2==0.21.0
-idna==3.4
+docutils==0.20.1
+exceptiongroup==1.2.0
+executing==2.0.1
+fastjsonschema==2.19.1
+fonttools==4.51.0
+google-api-core==2.18.0
+google-api-python-client==2.125.0
+google-auth==2.29.0
+google-auth-httplib2==0.2.0
+google-auth-oauthlib==1.2.0
+googleapis-common-protos==1.63.0
+httplib2==0.22.0
+idna==3.7
 imagesize==1.4.1
-importlib-metadata==6.0.0
-ipykernel==6.21.1
-ipython==8.10.0
-ipython-genutils==0.2.0
-ipywidgets==7.7.3
-jedi==0.18.2
-Jinja2==3.1.2
-jsonschema==3.2.0
-jupyter-book==0.13.0
-jupyter-cache==0.4.3
-jupyter-events==0.6.3
-jupyter-server-mathjax==0.2.6
-jupyter-sphinx==0.3.2
-jupyter_client==8.0.2
-jupyter_core==5.2.0
-jupyter_server==2.2.1
-jupyter_server_terminals==0.4.4
-jupyterlab-pygments==0.2.2
-jupyterlab-widgets==1.1.2
-kiwisolver==1.4.4
-latexcodec==2.0.1
-linkify-it-py==1.0.3
-lxml==4.9.2
-markdown-it-py==1.1.0
-MarkupSafe==2.1.2
-matplotlib==3.6.3
+importlib_metadata==7.1.0
+importlib_resources==6.4.0
+ipykernel==6.29.4
+ipython==8.18.1
+jedi==0.19.1
+Jinja2==3.1.3
+jsonschema==4.21.1
+jsonschema-specifications==2023.12.1
+jupyter-book==1.0.0
+jupyter-cache==1.0.0
+jupyter_client==8.6.1
+jupyter_core==5.7.2
+kiwisolver==1.4.5
+latexcodec==3.0.0
+linkify-it-py==2.0.3
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+matplotlib==3.8.4
 matplotlib-inline==0.1.6
-mdit-py-plugins==0.2.8
-mistune==0.8.4
-myst-nb==0.13.2
-myst-parser==0.15.2
-nbclassic==0.5.1
-nbclient==0.5.13
-nbconvert==6.5.4
-nbdime==3.1.1
-nbformat==5.7.3
-nest-asyncio==1.5.6
-notebook==6.5.2
-notebook_shim==0.2.2
-numpy==1.24.2
+mdit-py-plugins==0.4.0
+mdurl==0.1.2
+myst-nb==1.0.0
+myst-parser==2.0.0
+nbclient==0.10.0
+nbformat==5.10.4
+nest-asyncio==1.6.0
+numpy==1.26.4
 oauthlib==3.2.2
-packaging==23.0
-pandas==1.5.3
-pandocfilters==1.5.0
-parso==0.8.3
-pexpect==4.8.0
-pickleshare==0.7.5
-Pillow==9.4.0
-platformdirs==3.0.0
-prometheus-client==0.16.0
-prompt-toolkit==3.0.36
-proto-plus==1.22.2
-protobuf==4.21.12
-psutil==5.9.4
+packaging==24.0
+pandas==2.2.2
+parso==0.8.4
+pexpect==4.9.0
+pillow==10.3.0
+platformdirs==4.2.0
+prompt-toolkit==3.0.43
+proto-plus==1.23.0
+protobuf==4.25.3
+psutil==5.9.8
 ptyprocess==0.7.0
 pure-eval==0.2.2
-pyarrow==11.0.0
-pyasn1==0.4.8
-pyasn1-modules==0.2.8
+pyasn1==0.6.0
+pyasn1_modules==0.4.0
 pybtex==0.24.0
-pybtex-docutils==1.0.2
-pycparser==2.21
-pydata-sphinx-theme==0.8.1
-pyee==8.2.2
-Pygments==2.14.0
-pyparsing==3.0.9
-pyppeteer==1.0.2
-pyrsistent==0.19.3
-python-dateutil==2.8.2
-python-json-logger==2.0.5
-pytz==2022.7.1
-PyYAML==6.0
-pyzmq==25.0.0
-requests==2.28.2
-requests-oauthlib==1.3.1
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
+pybtex-docutils==1.0.3
+pydata-sphinx-theme==0.15.2
+pyee==11.1.0
+Pygments==2.17.2
+pyparsing==3.1.2
+pyppeteer==2.0.0
+python-dateutil==2.9.0.post0
+pytz==2024.1
+PyYAML==6.0.1
+pyzmq==25.1.2
+referencing==0.34.0
+requests==2.31.0
+requests-oauthlib==2.0.0
+rpds-py==0.18.0
 rsa==4.9
-Send2Trash==1.8.0
 six==1.16.0
-smmap==5.0.0
-sniffio==1.3.0
 snowballstemmer==2.2.0
-soupsieve==2.3.2.post1
-Sphinx==4.5.0
-sphinx-book-theme==0.3.3
+soupsieve==2.5
+Sphinx==7.2.6
+sphinx-book-theme==1.1.2
 sphinx-comments==0.0.3
-sphinx-copybutton==0.5.1
-sphinx-external-toc==0.2.4
-sphinx-jupyterbook-latex==0.4.7
+sphinx-copybutton==0.5.2
+sphinx-jupyterbook-latex==1.0.0
 sphinx-multitoc-numbering==0.1.3
-sphinx-thebe==0.1.2
+sphinx-thebe==0.3.1
 sphinx-togglebutton==0.3.2
-sphinx_design==0.1.0
-sphinxcontrib-applehelp==1.0.4
-sphinxcontrib-bibtex==2.5.0
-sphinxcontrib-devhelp==1.0.2
-sphinxcontrib-htmlhelp==2.0.1
+sphinx_design==0.5.0
+sphinx_external_toc==1.0.1
+sphinxcontrib-applehelp==1.0.8
+sphinxcontrib-bibtex==2.6.2
+sphinxcontrib-devhelp==1.0.6
+sphinxcontrib-htmlhelp==2.0.5
 sphinxcontrib-jsmath==1.0.1
-sphinxcontrib-qthelp==1.0.3
-sphinxcontrib-serializinghtml==1.1.5
-SQLAlchemy==1.4.46
-stack-data==0.6.2
-terminado==0.17.1
-tinycss2==1.2.1
-tornado==6.2
-tqdm==4.64.1
-traitlets==5.9.0
-uc-micro-py==1.0.1
+sphinxcontrib-qthelp==1.0.7
+sphinxcontrib-serializinghtml==1.1.10
+SQLAlchemy==2.0.29
+stack-data==0.6.3
+tabulate==0.9.0
+tornado==6.4
+tqdm==4.66.2
+traitlets==5.14.2
+typing_extensions==4.11.0
+tzdata==2024.1
+uc-micro-py==1.0.3
 uritemplate==4.1.1
-urllib3==1.26.14
-wcwidth==0.2.6
-webencodings==0.5.1
-websocket-client==1.5.1
+urllib3==1.26.18
+wcwidth==0.2.13
 websockets==10.4
-widgetsnbextension==3.6.2
-zipp==3.13.0
+zipp==3.18.1
